put the history and callback to keys into abstract class
